requirements.txt
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.46.2
datasets==3.1.0
tqdm==4.67.0
hydra-core==1.3.2
nltk==3.9.1
protobuf==5.28.3
tensorboardX==2.6.2.2
geoopt @ git+https://github.com/geoopt/geoopt@v0.5.1